# docs/swagger.yaml
basePath: /
definitions:
...
  model.ErrorResponse:
    properties:
      errors:
        type: string
    type: object
...
  model.SuccResp:
    properties:
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: API Avito PVZ Service
  version: 1.0.0
paths:
...
  /pvz/{pvzId}/close_last_reception:
    post:
      consumes:
      - application/json
      description: Закрытие последней открытой приемки товаров в рамках ПВЗ
      parameters:
      - description: ID ПВЗ
        in: path
        name: pvzId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Приемка закрыта
          schema:
            $ref: '#/definitions/model.SuccResp'
        "400":
          description: Неверный запрос или приемка уже закрыта
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Неавторизован.
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Доступ запрещен.
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера.
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Закрытие приемки
  /pvz/{pvzId}/delete_last_product:
    post:
      consumes:
      - application/json
      description: Удаление последнего добавленного товара из текущей приемки (LIFO,
        только для сотрудников ПВЗ)
      parameters:
      - description: ID ПВЗ
        in: path
        name: pvzId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Товар удален
          schema:
            $ref: '#/definitions/model.SuccResp'
        "400":
          description: Неверный запрос, нет активной приемки или нет товаров для удаления
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Неавторизован.
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "403":
          description: Доступ запрещен.
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера.
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление товара
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"